fluids: Add DMSetupByOrder_* helper function
