+++ examples/fluids/navierstokes.h
@@ -382,6 +382,11 @@ PetscErrorCode CreateDM(MPI_Comm comm, ProblemData *problem, MatType, VecType, D
 
 // Set up DM
 PetscErrorCode SetUpDM(DM dm, ProblemData *problem, PetscInt degree, PetscInt q_extra, SimpleBC bc, Physics phys);
+PetscErrorCode DMSetupByOrderBegin_FEM(PetscBool setup_faces, PetscBool setup_coords, PetscInt degree, PetscInt coord_order, PetscInt q_extra,
+                                       CeedInt num_fields, const CeedInt *field_sizes, DM dm);
+PetscErrorCode DMSetupByOrderEnd_FEM(PetscBool setup_coords, DM dm);
+PetscErrorCode DMSetupByOrder_FEM(PetscBool setup_faces, PetscBool setup_coords, PetscInt degree, PetscInt coord_order, PetscInt q_extra,
+                                  CeedInt num_fields, const CeedInt *field_sizes, DM dm);
 
 // Refine DM for high-order viz
 PetscErrorCode VizRefineDM(DM dm, User user, ProblemData *problem, SimpleBC bc, Physics phys);
